+++ home.work.2/src/main/java/ru/otus/service/PassExam.java
@@ -1,0 +1,8 @@
+package ru.otus.service;
+
+import ru.otus.domain.Student;
+
+public interface PassExam {
+
+    void passTestOrNot(int countCorrectPersonAnswers,int countCorrectPersonForPassTest, Student student);
+}
